refactor: Changed from float to double for more precise calculations.

+++ backend/lib/Kinematic/Kinematics.h
@@ -7,27 +7,26 @@
 
 struct DHparam
 {
-    float a;     ///< Link length
-    float alpha; ///< Link twist
-    float d;     ///< Link offset
-    // float theta; ///< Joint angle (dynamic parameter)
+    double a;     ///< Link length
+    double alpha; ///< Link twist
+    double d;     ///< Link offset
 };
 
 struct Pose
 {
-    float x, y, z;          ///< Position (mm)
-    float roll, pitch, yaw; ///< Orientation (Euler angles) in degrees
-    bool  inSingularity;    ///< Indicates if the pose is in a singularity
+    double x, y, z;          ///< Position (mm)
+    double roll, pitch, yaw; ///< Orientation (Euler angles) in degrees
+    bool   inSingularity;    ///< Indicates if the pose is in a singularity
 };
 
 struct Angles
 {
-    float theta1;
-    float theta2;
-    float theta3;
-    float theta4;
-    float theta5;
-    float theta6;
+    double theta1;
+    double theta2;
+    double theta3;
+    double theta4;
+    double theta5;
+    double theta6;
 };
 
 class Kinematics
@@ -37,25 +36,25 @@ public:
 
     Kinematics(const std::vector<MotorConfig>& motorConfigs, const std::vector<DHparam>& dhParams);
 
-    std::vector<float> getJointAnglesInRad() const;
-    void               setToolFrame(float x, float y, float z, float yaw, float pitch, float roll);
-    Pose               forwardKinematics();
-    Angles             inverseKinematics(float x, float y, float z, float yaw, float pitch, float roll);
+    std::vector<double> getJointAnglesInRad() const;
+    void                setToolFrame(double x, double y, double z, double yaw, double pitch, double roll);
+    Pose                forwardKinematics();
+    Angles              inverseKinematics(double x, double y, double z, double yaw, double pitch, double roll);
 
 private:
-    float _totalRatio(const MotorConfig& motorConfig) const;
-    float _stepsPerRev(const MotorConfig& motorConfig) const;
-    float _stepsToDeg(const MotorConfig& motorConfig, const int currPosInSteps) const;
-    int   _degToSteps(const MotorConfig& motorConfig, const float deg) const;
-    float _degToRad(const float deg) const;
-    float _radToDeg(const float rad) const;
+    double _totalRatio(const MotorConfig& motorConfig) const;
+    double _stepsPerRev(const MotorConfig& motorConfig) const;
+    double _stepsToDeg(const MotorConfig& motorConfig, const int currPosInSteps) const;
+    int    _degToSteps(const MotorConfig& motorConfig, const double deg) const;
+    double _degToRad(const double deg) const;
+    double _radToDeg(const double rad) const;
 
-    Eigen::Matrix4f _dhToTable(const DHparam& param, const float theta) const;
-    Eigen::Matrix4f _createTransformationMatrix(float x, float y, float z, float yaw, float pitch, float roll) const;
+    Eigen::Matrix4d _dhToTable(const DHparam& param, const double theta) const;
+    Eigen::Matrix4d _createTransformationMatrix(double x, double y, double z, double yaw, double pitch, double roll) const;
 
     const std::vector<MotorConfig> _motorConfigs;
     const std::vector<DHparam>     _dhParams;
-    Eigen::Matrix4f                _toolFrameMatrix = Eigen::Matrix4f::Identity();
+    Eigen::Matrix4d                _toolFrameMatrix = Eigen::Matrix4d::Identity();
 };
 
 #endif // KINEMATICS_H
